docs/source/Compute/Lambda-Root/Deploy-Lambda-and-Configure-Event-Mapping-with-AWS-CDK/sqs/requirements.txt
aws-cdk-lib==2.130.0
constructs>=10.0.0,<11.0.0
boto_session_manager>=1.7.2,<2.0.0
